Rewrite of USB comms to fix issue
USB comms to Tonex changed to now read the pedal state, and only update the preset indexes and send back. No longer overwriting the pedal configuration which is in the same data packet.
Also changed so that current preset is read at boot, rather than forcing it to preset 1.

Credit to Github user "vit3k" for great protocol analysis work.